refactor: split main into separate modules
