+++ YouTrackPlugin.ts
@@ -1,0 +1,249 @@
+import { Plugin, TFile, normalizePath, requestUrl } from "obsidian";
+import YouTrackSettingTab from "./YouTrackSettingTab";
+import YouTrackIssueModal from "./YouTrackIssueModal";
+interface YouTrackPluginSettings {
+	youtrackUrl: string;
+	apiToken: string;
+	useApiToken: boolean;
+	notesFolder: string;
+	templatePath: string;
+}
+
+const DEFAULT_SETTINGS: YouTrackPluginSettings = {
+	youtrackUrl: "https://youtrack.jetbrains.com",
+	apiToken: "",
+	useApiToken: false,
+	notesFolder: "YouTrack",
+	templatePath: "",
+};
+
+const DEFAULT_TEMPLATE = "# ${id}: ${title}\n\nURL: ${url}\n\n## Description\n\n${description}\n";
+
+const TIMESTAMP_FIELDS = new Set(["created", "updated", "resolved"]);
+
+interface DateTimeFormatOptions {
+	locale?: string;
+	timeZone?: string;
+}
+
+export default class YouTrackPlugin extends Plugin {
+	settings: YouTrackPluginSettings;
+	dateTimeOptions: DateTimeFormatOptions = {
+		locale: undefined, // Uses system locale by default
+		timeZone: undefined, // Uses system time zone by default
+	};
+
+	async onload() {
+		await this.loadSettings();
+
+		// Add settings tab
+		this.addSettingTab(new YouTrackSettingTab(this.app, this));
+
+		// Add command to fetch YouTrack issue
+		this.addCommand({
+			id: "fetch-youtrack-issue",
+			name: "Fetch YouTrack issue",
+			callback: () => {
+				// Open modal to input issue ID
+				new YouTrackIssueModal(this.app, this).open();
+			},
+		});
+
+		// Add ribbon icon
+		this.addRibbonIcon("clipboard-list", "Fetch YouTrack issue", () => {
+			new YouTrackIssueModal(this.app, this).open();
+		});
+	}
+
+	async loadSettings() {
+		this.settings = Object.assign({}, DEFAULT_SETTINGS, await this.loadData());
+	}
+
+	async saveSettings() {
+		await this.saveData(this.settings);
+	}
+
+	async readTemplateFile(): Promise<string | null> {
+		if (!this.settings.templatePath) {
+			return null;
+		}
+
+		try {
+			const normalizedPath = normalizePath(this.settings.templatePath);
+			const exists = await this.app.vault.adapter.exists(normalizedPath);
+
+			if (!exists) {
+				console.error(`Template file not found: ${normalizedPath}`);
+				return null;
+			}
+
+			return await this.app.vault.adapter.read(normalizedPath);
+		} catch (error) {
+			console.error("Error reading template file:", error);
+			return null;
+		}
+	}
+
+	async fetchIssueData(issueId: string): Promise<unknown> {
+		if (!this.settings.youtrackUrl) {
+			throw new Error("YouTrack URL is not set in plugin settings");
+		}
+
+		// Determine fields from the template
+		const template = (await this.readTemplateFile()) || DEFAULT_TEMPLATE;
+		const fieldList = this.parseFieldListFromTemplate(template);
+
+		const apiUrl = `${this.settings.youtrackUrl}/api/issues/${issueId}?fields=${fieldList.join(",")}`;
+
+		const headers: Record<string, string> = {
+			Accept: "application/json",
+			"Content-Type": "application/json",
+		};
+
+		// Add authorization if API token is set
+		if (this.settings.useApiToken && this.settings.apiToken) {
+			headers["Authorization"] = `Bearer ${this.settings.apiToken}`;
+		}
+
+		try {
+			const response = await requestUrl({
+				url: apiUrl,
+				method: "GET",
+				headers,
+			});
+
+			if (response.status != 200) {
+				throw new Error(`Error fetching issue: ${response.text} (${response.status})`);
+			}
+
+			return await response.json;
+		} catch (error) {
+			console.error("Error fetching YouTrack issue:", error);
+			throw error;
+		}
+	}
+
+	// Parse list of fields referenced in a template
+	parseFieldListFromTemplate(template: string): string[] {
+		const fields = new Set<string>();
+		const matches = template.matchAll(/\$\{([^}]+)\}/g);
+
+		for (const match of matches) {
+			const field = match[1].trim();
+			if (!field || field === "id" || field === "url") continue;
+
+			fields.add(field === "title" ? "summary" : field);
+		}
+
+		return Array.from(fields);
+	}
+
+	parseIssueId(input: string): string | null {
+		const trimmed = input.trim();
+		if (!trimmed) return null;
+
+		let id = trimmed;
+
+		if (/^https?:\/\//i.test(trimmed)) {
+			try {
+				const inputUrl = new URL(trimmed);
+				const baseUrl = new URL(this.settings.youtrackUrl);
+				const basePath = baseUrl.pathname.replace(/\/$/, "");
+				const prefix = `${basePath}/issue/`;
+
+				if (inputUrl.origin !== baseUrl.origin || !inputUrl.pathname.startsWith(prefix)) {
+					return null;
+				}
+
+				id = inputUrl.pathname.substring(prefix.length).split("/")[0];
+			} catch {
+				return null;
+			}
+		}
+
+		return /^[A-Za-z]+-\d+$/.test(id) ? id : null;
+	}
+	formatTimestamp(value: unknown): string {
+		const date = typeof value === "number" ? new Date(value) : new Date(String(value));
+		if (isNaN(date.getTime())) {
+			return String(value);
+		}
+
+		const formatOptions = this.dateTimeOptions;
+		return date.toLocaleString(formatOptions.locale, {
+			timeZone: formatOptions.timeZone,
+		});
+	}
+
+	renderTemplate(template: string, issueId: string, issueUrl: string, issueData: Record<string, unknown>): string {
+		// Build replacement map
+		const fieldList = this.parseFieldListFromTemplate(template);
+
+		// these fields are always replaced
+		const replacements: Record<string, string> = {
+			id: issueId,
+			url: issueUrl,
+		};
+
+		// always add "title" field from "summary" for backward compatibility
+		if (issueData.summary) {
+			replacements.title = String(issueData.summary);
+		}
+
+		// replace other fields specified in settings
+		for (const field of fieldList) {
+			const value = issueData[field];
+			if (value) {
+				let formatted = String(value);
+				if (TIMESTAMP_FIELDS.has(field)) {
+					formatted = this.formatTimestamp(value);
+				}
+				replacements[field] = formatted;
+			}
+		}
+		return template.replace(/\$\{([^}]+)\}/g, (_match, key) => replacements[key] ?? "");
+	}
+
+	async createIssueNote(issueId: string, issueData: Record<string, unknown>) {
+		// Create folder if it doesn't exist
+		const folderPath = this.settings.notesFolder ? this.settings.notesFolder : "";
+		if (folderPath) {
+			try {
+				const folderExists = await this.app.vault.adapter.exists(folderPath);
+				if (!folderExists) {
+					await this.app.vault.createFolder(folderPath);
+				}
+			} catch (error) {
+				console.error("Error creating folder:", error);
+				// Continue even if folder creation fails
+			}
+		}
+
+		// Create file name from issue ID
+		const fileName = `${folderPath ? folderPath + "/" : ""}${issueId}.md`;
+
+		// Construct issue URL
+		const issueUrl = `${this.settings.youtrackUrl}/issue/${issueId}`;
+
+		// Try to read template file
+		const template = (await this.readTemplateFile()) || DEFAULT_TEMPLATE;
+
+		// Create note content
+		const noteContent = this.renderTemplate(template, issueId, issueUrl, issueData);
+
+		// Create file in Obsidian vault
+		try {
+			const normalizedPath = normalizePath(fileName);
+			await this.app.vault.create(normalizedPath, noteContent);
+
+			// Open the newly created file
+			const file = this.app.vault.getAbstractFileByPath(normalizedPath);
+			if (file instanceof TFile) {
+				await this.app.workspace.getLeaf().openFile(file);
+			}
+		} catch (error) {
+			console.error("Error creating note:", error);
+			throw error;
+		}
+	}
+}
